move+hide+improve confidence interval computation
